+++ ktx-prefs-compiler/src/main/java/co/windly/ktxprefs/compiler/PrefsProcessor.kt
@@ -66,6 +66,8 @@ class PrefsProcessor : AbstractProcessor() {
     const val EDITOR_WRAPPER = "EditorWrapper"
 
     const val CONSTANTS = "Constants"
+
+    const val EXTENTION = "Ext"
   }
 
   //endregion
@@ -157,9 +159,11 @@ class PrefsProcessor : AbstractProcessor() {
           prefList += preference
         }
 
-        // Retrieve file name.
+        // Retrieve a class annotation.
         val prefsAnnotation = classElement.getAnnotation(Prefs::class.java)
-        val filename = when(prefsAnnotation.value.isEmpty()) {
+
+        // Retrieve file name.
+        val filename = when (prefsAnnotation.value.isEmpty()) {
           false -> prefsAnnotation.value
           true -> prefsAnnotation.fileName
         }
@@ -167,16 +171,27 @@ class PrefsProcessor : AbstractProcessor() {
         // Prepare arguments container.
         val arguments: MutableMap<String, Any?> = mutableMapOf()
         arguments["fileName"] = filename
+        arguments["distinctUntilChanged"] = prefsAnnotation.distinctUntilChanged
         arguments["package"] = packageElement.qualifiedName
         arguments["comment"] = classComment
         arguments["prefWrapperClassName"] = "${classElement.simpleName}${SuffixConfiguration.PREFS_WRAPPER}"
         arguments["editorWrapperClassName"] = "${classElement.simpleName}${SuffixConfiguration.EDITOR_WRAPPER}"
         arguments["constantsClassName"] = "${classElement.simpleName}${SuffixConfiguration.CONSTANTS}"
+        arguments["extensionClassName"] = "${classElement.simpleName}"
         arguments["prefList"] = prefList
 
         // Make directory for generated files.
         val packageDirectory = File(kaptTargetDirectory, getPackagePath(packageElement))
           .also { it.mkdirs() }
+
+        // Create shared preferences wrapper extensions.
+        File(packageDirectory, "${classElement.simpleName}${SuffixConfiguration.EXTENTION}.kt").apply {
+          writer(Charset.defaultCharset())
+            .use { writer ->
+              val template = freemarkerConfiguration.getTemplate(FreemarkerTemplate.EXTENSIONS)
+              template.process(arguments, writer)
+            }
+        }
 
         // Create shared preferences wrapper.
         File(packageDirectory, "${classElement.simpleName}${SuffixConfiguration.PREFS_WRAPPER}.kt").apply {
@@ -298,6 +313,8 @@ class PrefsProcessor : AbstractProcessor() {
   }
 
   private object FreemarkerTemplate {
+
+    const val EXTENSIONS = "extensions.ftl"
 
     const val PREFS_WRAPPER = "shared_preferences_wrapper.ftl"