+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Change Log
 
-### AndroidX Preferences 0.0.1 *(2019-02-03)*
+### Kotlin AndroidX Preferences 1.0.0 *(2019-02-21)*
+
+  * Add full annotation processor implementation.
+  * Add reactive extensions support.
+
+### Kotlin AndroidX Preferences 0.0.1 *(2019-02-03)*
 
  * Project scaffold.